requirements.txt
langchain>=0.2.12
langchain-community>=0.2.11
langchain-groq>=0.1.5
langchain-huggingface>=0.0.7
faiss-cpu>=1.8.0.post1
sentence-transformers>=3.0.1
unstructured>=0.15.7
unstructured[pdf]>=0.15.7
pypdf>=4.2.0
beautifulsoup4>=4.12.3
requests>=2.32.3
python-dotenv>=1.0.1
streamlit>=1.37.0
tiktoken>=0.7.0
ragas>=0.1.13
numpy>=1.26.4
pandas>=2.2.2
